[SwiftUI WebKit API] Expose _editable on WebPage https://bugs.webkit.org/show_bug.cgi?id=320772 rdar://183774945
Reviewed by Aditya Keerthi.

Expose SPI to define the content environment of a page, a more flexible version of the existing `_editable` WKWebView SPI.

Also refactor WebPage to have a dedicated SPI file.

* Source/WebKit/UIProcess/API/Swift/WebPage+SPI.swift: Added.
(WebPage.setMenuBuilder(_:)):
(WebPage.smartListsEnabled):
(terminateWebContentProcess):
(editorStateSnapshots):
(contentEnvironment):
* Source/WebKit/UIProcess/API/Swift/WebPage.swift:
(editorStateSnapshotsContinuations):
(setMenuBuilder(_:)): Deleted.
(smartListsEnabled): Deleted.
(terminateWebContentProcess): Deleted.
(editorStateSnapshots): Deleted.
* Source/WebKit/WebKit.xcodeproj/project.pbxproj:

Canonical link: https://commits.webkit.org/318398@main
